Rename "signals.py" to "indicators.py"
