+++ strategies/OscillatingStrategy.py
@@ -5,7 +5,7 @@ from pytz import timezone
 from typing import Union, Tuple, Sequence
 
 from analysis.financials import FinancialsMixin
-from models.signals import Signal, IndicatorContainer, Indicator
+from models.indicators import Signal, IndicatorContainer, Indicator
 from models.trades import Side
